+++ changelog.md
@@ -1,4 +1,4 @@
 # Changelog
 
 **./test-data/p1-preloaded_lists/referrals.sql**
-* Added test data for 'ReferralSourceCompany' table.
+* Defined 'registerTimestamp' for generic Referral Source entries. (2017-11-02 00:00:00)